+++ docs/analysis/spectral.rst
@@ -1,2 +1,6 @@
 Spectral
-============
+========
+
+.. autoclass:: oceanicospy.analysis.spectral.WaveSpectralAnalyzer
+   :members:
+   :undoc-members:
